+++ src/Symfony/Component/Workflow/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
  * Added context to the event dispatched
  * Dispatch an event when the subject enters in the workflow for the very first time
  * Added a default context to the previous event
+ * Added support for specifying which events should be dispatched when calling `workflow->apply()`
 
 5.1.0
 -----